+++ Unstable/StarFighter/StarFighter/Globals.h
@@ -5,17 +5,17 @@
 #define _USE_MATH_DEFINES
 #include <math.h>
 
-int RandomizeIntBetweenRatios(int value, float min_value, float max_value);
-float RandomizeFloatBetweenRatios(float value, float min_ratio, float max_ratio);
-float ProrataBetweenThreshold(float m_value, float min_threshold, float max_threshold);
+int RandomizeIntBetweenRatios(int value, double min_value, double max_value);
+double RandomizeFloatBetweenRatios(double value, double min_ratio, double max_ratio);
+double ProrataBetweenThreshold(double m_value, double min_threshold, double max_threshold);
 int RandomizeIntBetweenValues(int min_value, int max_value);
-int RandomizeIntBetweenFloats(float min_value, float max_value);
-float RandomizeFloatBetweenValues(float min_value, float max_value);
+int RandomizeIntBetweenFloats(double min_value, double max_value);
+double RandomizeFloatBetweenValues(double min_value, double max_value);
 int RandomizeSign();
-float MaxBetweenValues(float x, float y);
-float MinBetweenValues(float x, float y);
-float Lerp(float value, float input_min, float input_max, float output_min, float output_max);
-float Bound(float value, float output_min, float output_max);
-float CosInterpolation(float value, float input_min, float input_max, float output_min, float output_max);
+double MaxBetweenValues(double x, double y);
+double MinBetweenValues(double x, double y);
+double Lerp(double value, double input_min, double input_max, double output_min, double output_max);
+double Bound(double value, double output_min, double output_max);
+double CosInterpolation(double value, double input_min, double input_max, double output_min, double output_max);
 
 #endif // GLOBALS_H_INCLUDED
